firmware: Implement sysex messages for handling quantizer configs
Note: For testing purposes, these sysexes currently affect the live config
directly. A later commit will change them to affect the config in flash.

As the config struct is about 2KB in size, and we end up needing ~2 copies
on the stack in cmd_0x1A_read_quantizer_config_(), we need to increase the
stack size to 8KB to prevent overflow. This can likely be reduced in future.