Fix PWM and pitch "stepping" behavior
Previously, Gemini used the MCP2748 DAC's sequential update command to update the control voltages for amplitude compensation and pulse width. This command actually waits for the DAC's EEPROM to update before allowing another change so it significantly lowered refresh rate of the DAC (to around 25 ms!). This switches to using the "fast update" command instead which bypasses the EEPROM and theoretically allows updating at 166 kHz. However, because Gemini's update loop is only run once the ADC has new data, the effective update rate is 1.4 kHz. This is *plenty* fast enough and completely eliminates the stepping behavior.